# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Public. Opens a session for an existing account. The access token comes back in the body; the refresh token is set only as an `httpOnly`, path-scoped cookie and is never returned in a body. A wrong password and an unknown address answer the same 401, so the route cannot be used to probe which addresses have accounts.",
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session opened. Refresh cookie set on the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Wrong password, unknown address, or a disabled account.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Public. Creates the account and opens a session in one call. The access token comes back in the body; the refresh token is set only as an `httpOnly`, path-scoped cookie. The account may still need `POST /auth/complete-profile` before the rest of the API opens up — read `GET /auth/me` to find out.",
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created and session opened.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body, or a password that fails the policy.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "`CONFLICT` — That email is already registered.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "CONFLICT"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register an account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Public, but the refresh cookie is the credential — this route reads no body and accepts no fallback. Rotates the pair: the old refresh token is revoked, a new one replaces the cookie and a fresh access token comes back in the body. Presenting an access token here fails, and re-using an already-rotated refresh token is treated as replay.",
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pair rotated. New refresh cookie set on the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — No refresh cookie, or a token that is expired, revoked, replayed or of the wrong type.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rotate the session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Public, and deliberately forgiving: clears the refresh cookie and revokes the stored token when one is present, but answers 204 either way so a client can always reach a logged-out state.",
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Session closed and cookie cleared. Empty body."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log out",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "The account behind the bearer token, with its space and role — the context every other route scopes to. Reachable with an incomplete profile, so a client can read it to decide whether to send the user to `POST /auth/complete-profile` first.",
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current account, space and role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/complete-profile": {
      "post": {
        "description": "Fills in what registration or an invitation left open. One of the few routes reachable with an incomplete profile — every other protected route answers 403 until this succeeds. Re-issues the session afterwards, because the token claims change: a new access token comes back in the body and the refresh cookie is replaced.",
        "operationId": "AuthController_completeProfile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile completed and session re-issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "`CONFLICT` — The profile is already completed.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "CONFLICT"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complete the profile",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tu Seguridad API",
    "description": "Space-scoped video surveillance API. Every route outside `auth` and `health` resolves resources inside the caller space, taken from the bearer token, so an id from another space answers 404 rather than a result. Writes that change what the system watches are admin-only. Failures answer `{ statusCode, code, message }` — branch on `code`, not on the message.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404,
            "description": "HTTP status, the same value as the response status line."
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "INVALID_ZONE",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "UPSTREAM_ERROR",
              "UPSTREAM_TIMEOUT",
              "INTERNAL_ERROR"
            ],
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code. Branch on this, not on the message."
          },
          "message": {
            "type": "string",
            "example": "Camera 7f3a not found",
            "description": "Human-readable reason. Never carries credentials, tokens or snapshot URLs."
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "example": "change-me"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AccessTokenDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Bearer token for the `Authorization` header. Short-lived — rotate through `POST /auth/refresh` when it expires.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 12
          },
          "firstName": {
            "type": "string",
            "example": "Ada"
          },
          "lastName": {
            "type": "string",
            "example": "Lovelace"
          },
          "phone": {
            "type": "string",
            "example": "+5491122334455",
            "description": "E.164"
          },
          "spaceName": {
            "type": "string",
            "example": "My Secure Space"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "phone",
          "spaceName"
        ]
      },
      "MeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string",
            "example": "owner@example.com"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "+5491122334455"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "profileCompleted": {
            "type": "boolean",
            "description": "False until an invited member completes name, phone and password"
          },
          "spaceId": {
            "type": "string",
            "format": "uuid"
          },
          "spaceName": {
            "type": "string",
            "example": "My Secure Space"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "member"
            ],
            "example": "admin"
          },
          "receiveAlerts": {
            "type": "boolean",
            "description": "Whether this member receives alert deliveries"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "phone",
          "isActive",
          "profileCompleted",
          "spaceId",
          "spaceName",
          "role",
          "receiveAlerts"
        ]
      },
      "CompleteProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Ada"
          },
          "lastName": {
            "type": "string",
            "example": "Lovelace"
          },
          "phone": {
            "type": "string",
            "example": "+5491122334455",
            "description": "E.164"
          },
          "password": {
            "type": "string",
            "minLength": 12
          },
          "avatarUrl": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "phone",
          "password"
        ]
      },
...
    }
  }
}